+++ tests/LinqContraband.Tests/Analyzers/LC011_EntityMissingPrimaryKey/EntityMissingPrimaryKeyEdgeCasesTests.cs
@@ -62,6 +62,60 @@ namespace TestNamespace
     }
 }";
 
+    private const string SemanticMockAttributes = @"
+namespace System.ComponentModel.DataAnnotations
+{
+    public class KeyAttribute : Attribute {}
+}
+namespace System.ComponentModel.DataAnnotations.Schema
+{
+    public class NotMappedAttribute : Attribute {}
+}
+namespace Microsoft.EntityFrameworkCore
+{
+    public class KeylessAttribute : Attribute {}
+    public class PrimaryKeyAttribute : Attribute
+    {
+        public PrimaryKeyAttribute(params string[] propertyNames) {}
+    }
+    public class DbContext : IDisposable
+    {
+        public void Dispose() {}
+        protected virtual void OnModelCreating(ModelBuilder modelBuilder) {}
+    }
+    public class DbSet<T> where T : class {}
+
+    public interface IEntityTypeConfiguration<T> where T : class
+    {
+        void Configure(EntityTypeBuilder<T> builder);
+    }
+
+    public class ModelBuilder
+    {
+        public EntityTypeBuilder<T> Entity<T>() where T : class => new EntityTypeBuilder<T>();
+        public void ApplyConfiguration<T>(IEntityTypeConfiguration<T> configuration) where T : class {}
+        public void ApplyConfigurationsFromAssembly(System.Reflection.Assembly assembly) {}
+    }
+
+    public class EntityTypeBuilder<T> where T : class
+    {
+        public EntityTypeBuilder<T> HasKey(params string[] propertyNames) => this;
+        public EntityTypeBuilder<T> HasKey(System.Linq.Expressions.Expression<Func<T, object>> keyExpression) => this;
+        public EntityTypeBuilder<T> HasNoKey() => this;
+        public EntityTypeBuilder<T> ToTable(string name) => this;
+        public OwnedNavigationBuilder<T, TOwned> OwnsOne<TOwned>(System.Linq.Expressions.Expression<Func<T, TOwned>> navigationExpression) where TOwned : class => new OwnedNavigationBuilder<T, TOwned>();
+        public OwnedNavigationBuilder<T, TOwned> OwnsMany<TOwned>(System.Linq.Expressions.Expression<Func<T, System.Collections.Generic.IEnumerable<TOwned>>> navigationExpression) where TOwned : class => new OwnedNavigationBuilder<T, TOwned>();
+    }
+
+    public class OwnedNavigationBuilder<T, TOwned> where T : class where TOwned : class {}
+}
+
+namespace TestNamespace
+{
+    public class MyDbContext : Microsoft.EntityFrameworkCore.DbContext
+    {
+";
+
     [Fact]
     public async Task TestInnocent_EntityInheritsIdFromBaseClass_ShouldNotTrigger()
     {
@@ -303,6 +357,332 @@ namespace TestNamespace
     {
         public string Value { get; set; }
         public int Count { get; set; }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestCrime_FakeAttributeNames_ShouldNotSuppressDiagnostic()
+    {
+        var test = @"
+using System;
+using FakeAnnotations;
+using Microsoft.EntityFrameworkCore;
+
+namespace FakeAnnotations
+{
+    public class KeyAttribute : Attribute {}
+    public class KeylessAttribute : Attribute {}
+    public class PrimaryKeyAttribute : Attribute
+    {
+        public PrimaryKeyAttribute(params string[] propertyNames) {}
+    }
+}
+
+namespace Microsoft.EntityFrameworkCore
+{
+    public class DbContext : IDisposable
+    {
+        public void Dispose() {}
+    }
+
+    public class DbSet<T> where T : class {}
+}
+
+namespace TestNamespace
+{
+    public class MyDbContext : Microsoft.EntityFrameworkCore.DbContext
+    {
+        public DbSet<FakeKeyEntity> {|LC011:FakeKeys|} { get; set; }
+        public DbSet<FakeKeylessEntity> {|LC011:FakeKeyless|} { get; set; }
+        public DbSet<FakePrimaryKeyEntity> {|LC011:FakePrimaryKeys|} { get; set; }
+    }
+
+    public class FakeKeyEntity
+    {
+        [Key]
+        public int Code { get; set; }
+    }
+
+    [Keyless]
+    public class FakeKeylessEntity
+    {
+        public string Name { get; set; }
+    }
+
+    [PrimaryKey(nameof(Code))]
+    public class FakePrimaryKeyEntity
+    {
+        public int Code { get; set; }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestCrime_NotMappedId_ShouldTrigger()
+    {
+        var test = Usings + SemanticMockAttributes + @"
+        public DbSet<IgnoredIdEntity> {|LC011:IgnoredIds|} { get; set; }
+    }
+
+    public class IgnoredIdEntity
+    {
+        [NotMapped]
+        public int Id { get; set; }
+        public string Name { get; set; }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestCrime_UnappliedEntityTypeConfiguration_ShouldTrigger()
+    {
+        var test = Usings + SemanticMockAttributes + @"
+        public DbSet<ConfiguredElsewhereEntity> {|LC011:ConfiguredElsewhere|} { get; set; }
+    }
+
+    public class ConfiguredElsewhereEntity
+    {
+        public int Code { get; set; }
+    }
+
+    public class ConfiguredElsewhereEntityConfiguration : IEntityTypeConfiguration<ConfiguredElsewhereEntity>
+    {
+        public void Configure(EntityTypeBuilder<ConfiguredElsewhereEntity> builder)
+        {
+            builder.HasKey(e => e.Code);
+        }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestInnocent_AppliedEntityTypeConfiguration_ShouldNotTrigger()
+    {
+        var test = Usings + SemanticMockAttributes + @"
+        public DbSet<AppliedConfigEntity> AppliedConfigs { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.ApplyConfiguration(new AppliedConfigEntityConfiguration());
+        }
+    }
+
+    public class AppliedConfigEntity
+    {
+        public int Code { get; set; }
+    }
+
+    public class AppliedConfigEntityConfiguration : IEntityTypeConfiguration<AppliedConfigEntity>
+    {
+        public void Configure(EntityTypeBuilder<AppliedConfigEntity> builder)
+        {
+            builder.HasKey(e => e.Code);
+        }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestInnocent_BuilderVariableHasKey_ShouldNotTrigger()
+    {
+        var test = Usings + SemanticMockAttributes + @"
+        public DbSet<VariableConfiguredEntity> VariableConfiguredEntities { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            var entity = modelBuilder.Entity<VariableConfiguredEntity>();
+            entity.HasKey(e => e.Code);
+        }
+    }
+
+    public class VariableConfiguredEntity
+    {
+        public int Code { get; set; }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestInnocent_ChainedBuilderVariableHasKey_ShouldNotTrigger()
+    {
+        var test = Usings + SemanticMockAttributes + @"
+        public DbSet<ChainedConfiguredEntity> ChainedConfiguredEntities { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            var entity = modelBuilder.Entity<ChainedConfiguredEntity>();
+            entity.ToTable(""ChainedConfiguredEntities"").HasKey(e => e.Code);
+        }
+    }
+
+    public class ChainedConfiguredEntity
+    {
+        public int Code { get; set; }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestInnocent_AppliedConfigurationChainedBuilderHasKey_ShouldNotTrigger()
+    {
+        var test = Usings + SemanticMockAttributes + @"
+        public DbSet<ChainedConfigEntity> ChainedConfigEntities { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.ApplyConfiguration(new ChainedConfigEntityConfiguration());
+        }
+    }
+
+    public class ChainedConfigEntity
+    {
+        public int Code { get; set; }
+    }
+
+    public class ChainedConfigEntityConfiguration : IEntityTypeConfiguration<ChainedConfigEntity>
+    {
+        public void Configure(EntityTypeBuilder<ChainedConfigEntity> builder)
+        {
+            builder.ToTable(""ChainedConfigEntities"").HasKey(e => e.Code);
+        }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestInnocent_LocalVariableAppliedConfiguration_ShouldNotTrigger()
+    {
+        var test = Usings + SemanticMockAttributes + @"
+        public DbSet<VariableAppliedConfigEntity> VariableAppliedConfigEntities { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            var config = new VariableAppliedConfigEntityConfiguration();
+            modelBuilder.ApplyConfiguration(config);
+        }
+    }
+
+    public class VariableAppliedConfigEntity
+    {
+        public int Code { get; set; }
+    }
+
+    public class VariableAppliedConfigEntityConfiguration : IEntityTypeConfiguration<VariableAppliedConfigEntity>
+    {
+        public void Configure(EntityTypeBuilder<VariableAppliedConfigEntity> builder)
+        {
+            builder.HasKey(e => e.Code);
+        }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestCrime_ExternalApplyConfigurationsFromAssembly_ShouldNotApplyLocalConfig()
+    {
+        var test = Usings + SemanticMockAttributes + @"
+        public DbSet<ExternalAssemblyConfigEntity> {|LC011:ExternalAssemblyConfigs|} { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.ApplyConfigurationsFromAssembly(typeof(string).Assembly);
+        }
+    }
+
+    public class ExternalAssemblyConfigEntity
+    {
+        public int Code { get; set; }
+    }
+
+    public class ExternalAssemblyConfigEntityConfiguration : IEntityTypeConfiguration<ExternalAssemblyConfigEntity>
+    {
+        public void Configure(EntityTypeBuilder<ExternalAssemblyConfigEntity> builder)
+        {
+            builder.HasKey(e => e.Code);
+        }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestInnocent_ScopedBuilderVariableReuse_ShouldNotTrigger()
+    {
+        var test = Usings + SemanticMockAttributes + @"
+        public DbSet<ScopedOrder> ScopedOrders { get; set; }
+        public DbSet<ScopedCustomer> ScopedCustomers { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            {
+                var entity = modelBuilder.Entity<ScopedOrder>();
+                entity.HasKey(e => e.Code);
+            }
+
+            {
+                var entity = modelBuilder.Entity<ScopedCustomer>();
+                entity.HasKey(e => e.Code);
+            }
+        }
+    }
+
+    public class ScopedOrder
+    {
+        public int Code { get; set; }
+    }
+
+    public class ScopedCustomer
+    {
+        public int Code { get; set; }
+    }
+}";
+
+        await VerifyCS.VerifyAnalyzerAsync(test);
+    }
+
+    [Fact]
+    public async Task TestInnocent_NonGenericOwnsOne_ShouldNotTriggerForOwnedDbSetType()
+    {
+        var test = Usings + SemanticMockAttributes + @"
+        public DbSet<Order> Orders { get; set; }
+        public DbSet<ShippingAddress> ShippingAddresses { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.Entity<Order>().OwnsOne(o => o.ShippingAddress);
+        }
+    }
+
+    public class Order
+    {
+        public int Id { get; set; }
+        public ShippingAddress ShippingAddress { get; set; }
+    }
+
+    public class ShippingAddress
+    {
+        public string Street { get; set; }
+        public string City { get; set; }
     }
 }";